+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0] - 2018-02-07
+
+## Added
+
+- Signals ('onAccept', 'onReject', 'onError' & 'onFinal')
+- New option (boolean) 'throw_errors' in middleware.
+
 ## [1.0.1] - 2018-01-16
 
 ### Added
@@ -46,6 +53,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial dev release. Still unstable.
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/revam/git-koa-smart-proxy/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/revam/git-koa-smart-proxy/compare/v1.0.1..v1.1.0
 [1.0.1]: https://github.com/revam/git-koa-smart-proxy/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/revam/git-koa-smart-proxy/compare/v0.0.1...v1.0.0